+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [1.4.5](https://github.com/deltaDNA/ios-smartads-sdk/releases/tag/1.4.5) (????-??-??)
 ### Fixed
 - Not instantiating Facebook adapter.
+- AdColony reward callback ordering.
 
 ## [1.4.4](https://github.com/deltaDNA/ios-smartads-sdk/releases/tag/1.4.4) (2017-06-08)
 ### Fixed